Moved to named import and with to styled API
