+++ src/components/Dashboard/ThisWeek/LateCommitments/LateCommitments.test.tsx
@@ -1,6 +1,6 @@
 import React from 'react';
 import { render } from '../../../../../__tests__/util/testingLibraryReactMock';
-import LateCommitments from '.';
+import LateCommitments from './LateCommitments';
 
 describe('LateCommitments', () => {
   it('default', () => {
