--- src/components/Dashboard/ThisWeek/LateCommitments/index.tsx
@@ -1,3 +1,0 @@
-import LateCommitments from './LateCommitments';
-
-export default LateCommitments;
